Update option name and visual test cases
