+++ examples/next/visual-tests/js/demo/src/demos/textEllipsis/data.js
