+++ examples/next/visual-tests/js/demo/src/demos/textEllipsis/index.js
@@ -1,0 +1,24 @@
+import Handsontable from "handsontable";
+import { getDirectionFromURL, getThemeNameFromURL, getFromURL } from "../../utils";
+import { data } from './data';
+
+const root = document.getElementById('root');
+const example = document.createElement('div');
+
+root.appendChild(example);
+
+export function init() {
+  new Handsontable(example, {
+    data,
+    layoutDirection: getDirectionFromURL(),
+    language: getDirectionFromURL() === "rtl" ? arAR.languageCode : "en-US",
+    themeName: getThemeNameFromURL(),
+    colWidths: 100,
+    rowHeaders: true,
+    colHeaders: true,
+    textEllipsis: true,
+    licenseKey: "non-commercial-and-evaluation",
+  });
+
+  console.log(`Handsontable: v${Handsontable.version} (${Handsontable.buildDate})`);
+}
