+++ handsontable/src/__tests__/core/settings.types.ts
@@ -215,7 +215,7 @@ const allSettings: Required<Handsontable.GridSettings> = {
   strict: true,
   tableClassName: oneOf('foo', ['first-class-name', 'second-class-name']),
   tabMoves: oneOf({ col: 1, row: 1 }, (event: KeyboardEvent) => ({row: 2, col: 2})),
-  textTruncate: false,
+  textEllipsis: false,
   themeName: 'ht-theme-some-theme',
   title: 'foo',
   trimDropdown: true,
