+++ handsontable/src/renderers/baseRenderer/baseRenderer.js
@@ -10,7 +10,7 @@ import {
 import { A11Y_INVALID, A11Y_READONLY } from '../../helpers/a11y';
 
 export const RENDERER_TYPE = 'base';
-const TEXT_TRUNCATE_CLASS_NAME = 'htTextTruncate';
+const TEXT_ELLIPSIS_CLASS_NAME = 'htTextEllipsis';
 
 /**
  * @param {Core} hotInstance The Handsontable instance.
@@ -66,8 +66,8 @@ export function baseRenderer(hotInstance, TD, row, col, prop, value, cellPropert
     classesToAdd.push(cellProperties.placeholderCellClassName);
   }
 
-  if (cellProperties.textTruncate) {
-    classesToAdd.push(TEXT_TRUNCATE_CLASS_NAME);
+  if (cellProperties.textEllipsis) {
+    classesToAdd.push(TEXT_ELLIPSIS_CLASS_NAME);
   }
 
   removeClass(TD, classesToRemove);
