+++ handsontable/src/styles/base/_base.scss
@@ -440,7 +440,7 @@
       z-index: -1;
     }
 
-    .htTextTruncate {
+    .htTextEllipsis {
       text-overflow: ellipsis;
       overflow: hidden;
       white-space: nowrap;
