--- visual-tests/tests/js-only/text-truncate/text-truncate-column.spec.ts
@@ -1,17 +1,0 @@
-import { test } from '../../../src/test-runner';
-import { helpers } from '../../../src/helpers';
-
-test.skip(helpers.hotWrapper !== 'js', 'This test case is only for JavaScript framework');
-
-test(__filename, async({ goto, tablePage }) => {
-  await goto(
-    helpers
-      .setBaseUrl('/text-truncate-demo')
-      .setPageParams({
-        truncateTextOption: 'column',
-      })
-      .getFullUrl()
-  );
-
-  await tablePage.screenshot({ path: helpers.screenshotPath() });
-});
